+++ scripts/settings.js
@@ -1,4 +1,44 @@
 console.log("%cPF2e Alchemist Remaster Duct Tape: settings.js loaded","color: aqua; font-weight: bold;");
+// Function for debugging
+export function debugLog(logMsg, logType = "c", logLevel = "1") {
+	const debugEnabled = game.settings.get("pf2e-alchemist-remaster-ducttape", "debugEnabled");
+	if (!debugEnabled && logLevel != 3) return;
+		switch (logType) {
+			case "c": //console
+				switch (logLevel) {
+					case "1": // info/log
+						console.log(`%cP2Fe Alchemist Duct Tape (QuickAlchemy.js) | ${logMsg}`,"color: aqua; font-weight: bold;");
+						break;
+					case "2": // warn
+						console.warn(`P2Fe Alchemist Duct Tape (QuickAlchemy.js) | ${logMsg}`);
+						break;
+					case "3": // error
+						console.error(`P2Fe Alchemist Duct Tape (QuickAlchemy.js) | ${logMsg}`);
+						break;
+					default:
+						console.log(`%cP2Fe Alchemist Duct Tape (QuickAlchemy.js) | ${logMsg}`,"color: aqua; font-weight: bold;");
+				}
+				break;
+			case "u": // ui
+				switch (logLevel) {
+					case "1": // info/log
+						ui.notifications.info(`Alchemist Duct Tape (QuickAlchemy.js) | ${logMsg}`);
+						break;
+					case "2": // warn
+						ui.notifications.warn(`Alchemist Duct Tape (QuickAlchemy.js) | ${logMsg}`);
+						break;
+					case "3": // error
+						ui.notifications.error(`Alchemist Duct Tape (QuickAlchemy.js) | ${logMsg}`);
+						break;
+					default:
+						ui.notifications.info(logMsg);
+				}
+				break;
+			default:
+				console.warn(`P2Fe Alchemist Duct Tape (QuickAlchemy.js) | Invalid log event.`);
+		}
+}
+
 Hooks.once("init", () => {
 	
 	/* 
@@ -58,6 +98,15 @@ Hooks.once("init", () => {
 		requiresReload: true
 	});
 	
+	game.settings.register("pf2e-alchemist-remaster-ducttape", "showQACounts", {
+		name: "Quick Alchemy: Show forumla counts",
+		hint: "If enabled, will show the number of available formulas.",
+		scope: "world", // "client" or "world" depending on your use case
+		config: true,    // Whether to show this in the module settings UI
+		default: false,  // The default value of the setting
+		type: Boolean,   // The type of setting (true/false)
+		requiresReload: true,
+	});
 	
 	
 	/*